backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0,<5.0.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic[email]
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
tavily-python>=0.3.0
Pillow>=10.0.0
cryptography>=42.0.0
